Rename files & remove former implementation
