--- src/capture-dom.js.flow
@@ -1,10 +1,0 @@
-// @flow strict
-
-import type { Resource } from "./archiver.js"
-
-export type Options = {
-  docUrl?:string;
-  getDocInFrame?:HTMLIFrameElement => ?Document
-}
-
-declare export default (doc:Document, options?:Options) => Resource
