+++ src/resource.js
